pyproject.toml
[tool.poetry]
name = "imoveiswebscrap"
version = "0.1.0"
description = ""
authors = ["Gustavo Miranda <gustavocataldomiranda@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
scrapy = "^2.11.2"
scrapy-playwright = "^0.0.41"
pandas = "^2.2.3"
psycopg2-binary = "^2.9.9"
python-dotenv = "^1.0.1"
sqlalchemy = "^2.0.35"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
